chore: cleanup en_US/src/acronyms.txt from duplicates
These acronyms are already included in the en_shared/src/acronyms.txt
So we don't need them in this en_US/src/acronyms.txt file.

There is no need to do that normally, but I want to clean up and
identify the acronyms that are only in the en_US dictionary for now.