feat: add en_shared/shared-additional-words-ize
This source is specific to some English dictionaries:
 - American English (en_US)
 - British English (en_GB)

These words are not used in other English dictionaries, such as:
 - Australian English (en_AU)
 - Canadian English (en_CA)
 - British (-ise only) (en_GB-ise).

They could be present in them or not.
This list is not about adding or excluding them from these dictionaries.
The goal is to add them to en_GB and en_US dictionaries only.

This commit is only about setting up things.
The words will be added in future pull requests.