chore: rename software-terms.txt to software-terms-legacy.txt
This file should not be used anymore.
Its content comes from years of contributions.
It no longer reflects the expected content of the software-terms.txt file.
Let's consider all its content as legacy.

New PRs might be created to move some of the content to a new software-terms.txt file.